Reimplement is_val subset, fix val subset params
Always fix val subset params as follows, we don't want any changes happening! We want the ruler (val subset) to remain the same each time we measure (calculate validation loss).

Be aware val loss will shift regardless of changing any settings as a result, this is not an issue or due to any change in the underlying training process.